Domain Entities'de ayarlamalar yapıldı.

+++ FuarProjesi/Models/Entities/FoodService.cs
@@ -3,7 +3,7 @@
     public class FoodService : BaseEntity
     {
         public string ServiceName { get; set; }
-        public int ServicePrice { get; set; }
+        public decimal ServicePrice { get; set; }
 
         //Relational Property
         public virtual ICollection<Package> Packages { get; set; }
